+++ packages/gatsby-plugin-postcss/src/__tests__/gatsby-node.js
@@ -20,6 +20,14 @@ describe(`gatsby-plugin-postcss`, () => {
       "No options": {},
       "PostCss options": { postCssPlugins: [`autoprefixer`], sourceMap: false },
       "Css options": { cssLoaderOptions: { camelCase: true } },
+      "Css options commonjs": {
+        cssLoaderOptions: {
+          esModule: false,
+          modules: {
+            namedExport: false,
+          },
+        },
+      },
     },
     configs: {
       "No options": jest.fn().mockReturnValue({
@@ -76,11 +84,34 @@ describe(`gatsby-plugin-postcss`, () => {
           ],
         },
       }),
+      "Css options commonjs": jest.fn().mockReturnValue({
+        module: {
+          rules: [
+            {
+              oneOf: [
+                {
+                  test: /\.css$/,
+                  loaders: [`css-loader`],
+                },
+                {
+                  test: /\.module\.css$/,
+                  loaders: [`css-loader`],
+                },
+              ],
+            },
+            {
+              test: /\.js/,
+              use: [`babel-loader`],
+            },
+          ],
+        },
+      }),
     },
     actions: {
       "No options": actions.setWebpackConfig,
       "PostCss options": actions.replaceWebpackConfig,
       "Css options": actions.replaceWebpackConfig,
+      "Css options commonjs": actions.replaceWebpackConfig,
     },
   }