+++ packages/gatsby-source-contentful/src/gatsby-node.js
@@ -45,7 +45,7 @@ const validateContentfulAccess = async pluginOptions => {
           pluginOptions.spaceId
         )}" on environment "${
           pluginOptions.environment
-        } with access token "${maskText(
+        }" with access token "${maskText(
           pluginOptions.accessToken
         )}". Make sure to double check them!`
 
@@ -105,7 +105,7 @@ List of locales and their codes can be found in Contentful app -> Settings -> Lo
         .description(
           `Number of entries to retrieve from Contentful at a time. Due to some technical limitations, the response payload should not be greater than 7MB when pulling content from Contentful. If you encounter this issue you can set this param to a lower number than 100, e.g 50.`
         )
-        .default(100),
+        .default(1000),
       assetDownloadWorkers: Joi.number()
         .integer()
         .description(
@@ -134,6 +134,14 @@ List of locales and their codes can be found in Contentful app -> Settings -> Lo
     If you are confident your Content Types will have natural-language IDs (e.g. \`blogPost\`), then you should set this option to \`false\`. If you are unable to ensure this, then you should leave this option set to \`true\` (the default).`
         )
         .default(true),
+      contentfulClientConfig: Joi.object()
+        .description(
+          `Additional config which will get passed to [Contentfuls JS SDK](https://github.com/contentful/contentful.js#configuration).
+
+          Use this with caution, you might override values this plugin does set for you to connect to Contentful.`
+        )
+        .unknown(true)
+        .default({}),
       // default plugins passed by gatsby
       plugins: Joi.array(),
     })