+++ packages/gatsby-transformer-pdf/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.3.0-next.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-pdf@2.2.0-next.0...gatsby-transformer-pdf@2.3.0-next.0) (2021-03-25)
+
+**Note:** Version bump only for package gatsby-transformer-pdf
+
 # [2.2.0-next.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-pdf@2.1.0-next.2...gatsby-transformer-pdf@2.2.0-next.0) (2021-03-11)
 
 **Note:** Version bump only for package gatsby-transformer-pdf